+++ pylabrobot/only_fans/__init__.py
@@ -1,3 +1,3 @@
 from .backend import FanBackend
 from .fan import Fan
-from .hamilton_hepa_fan import HamiltonHepaFanBackend
+from .hamilton_hepa_fan_backend import HamiltonHepaFanBackend
